refactor: crate::{member,problem,solution}
