+++ src/main.rs
@@ -1,3 +1,10 @@
+use crate::problem::ProxyProblem;
+use crate::solution::ProxySolution;
+
+pub mod member;
+pub mod problem;
+pub mod solution;
+
 #[macro_use]
 extern crate derive_new;
 extern crate matchmaker;
@@ -7,159 +14,22 @@ extern crate rocket;
 extern crate rocket_slogger;
 #[macro_use]
 extern crate slog_derive;
-use indexmap::map::IndexMap;
-use indexmap::set::IndexSet;
-use matchmaker::da_stb::match_students;
-use matchmaker::{Category, Student};
-use rand::{rngs::StdRng, SeedableRng};
+
 use rocket::config::Config;
 use rocket::http::Status;
 use rocket::log::LogLevel;
 use rocket::serde::json::{json, Json, Value};
-use rocket::serde::{Deserialize, Serialize};
+
 use rocket_slogger::Slogger;
 
 #[get("/health/ready")]
-fn health_ready(log: Slogger) -> Status {
+fn get_health_ready(log: Slogger) -> Status {
     info!(log, "Ready");
     Status::NoContent
 }
 
-type MemberId = String;
-
-#[derive(Deserialize)]
-struct MemberInfo {
-    id: MemberId,
-    preferences: Vec<MemberId>,
-}
-
-trait Member {
-    fn from_info(info: &MemberInfo, present: &IndexMap<MemberId, Category>) -> Student;
-}
-
-impl Member for Student {
-    fn from_info(info: &MemberInfo, present: &IndexMap<MemberId, Category>) -> Student {
-        Student {
-            name: info.id.clone(),
-            preferences: info
-                .preferences
-                .iter()
-                .filter(|&k| present.contains_key(k))
-                .map(|k| present.get(k).unwrap())
-                .cloned()
-                .collect(),
-            exclude: present
-                .values()
-                .filter(|&v| !info.preferences.contains(&v.name))
-                .map(|v| present.get(&v.name).unwrap())
-                .cloned()
-                .collect(),
-        }
-    }
-}
-
-#[derive(Deserialize)]
-struct ProxyProblem {
-    capacity: usize,
-    members: Vec<MemberInfo>,
-    members_present: Vec<MemberId>,
-}
-
-#[derive(new, Serialize)]
-struct ProxySolution {
-    #[serde(skip)]
-    capacity: usize,
-
-    #[new(default)]
-    #[serde(skip)]
-    members_present: IndexMap<MemberId, Category>,
-
-    #[new(default)]
-    #[serde(skip)]
-    members_absent: IndexMap<MemberId, Student>,
-
-    #[new(default)]
-    members_represented: IndexMap<MemberId, MemberId>,
-
-    #[new(default)]
-    members_unrepresented: IndexSet<MemberId>,
-}
-
-#[derive(Clone, SerdeValue, Serialize)]
-struct ProxyMetrics {
-    capacity: usize,
-    total: usize,
-    present: usize,
-    absent: usize,
-    represented: usize,
-    unrepresented: usize,
-}
-
-impl ProxySolution {
-    fn from_problem(problem: &ProxyProblem) -> Self {
-        let mut solution = Self::new(problem.capacity);
-        solution.load_attendance(&problem.members_present);
-        solution.load_preferences(&problem.members);
-
-        solution
-    }
-
-    fn get_metrics(&mut self) -> ProxyMetrics {
-        ProxyMetrics {
-            capacity: self.capacity,
-            total: self.members_present.len() + self.members_absent.len(),
-            present: self.members_present.len(),
-            absent: self.members_absent.len(),
-            represented: self.members_represented.len(),
-            unrepresented: self.members_unrepresented.len(),
-        }
-    }
-
-    fn load_attendance(&mut self, members_present: &Vec<MemberId>) {
-        for id in members_present {
-            let present = Category::new(id, self.capacity);
-            self.members_present.insert(id.clone(), present);
-        }
-    }
-
-    fn load_preferences(&mut self, members: &Vec<MemberInfo>) {
-        for info in members {
-            if !self.members_present.contains_key(&info.id) {
-                let absent = <Student as Member>::from_info(info, &self.members_present);
-                self.members_absent.insert(info.id.clone(), absent);
-                self.members_unrepresented.insert(info.id.clone());
-            }
-        }
-    }
-
-    fn solve(&mut self, log: &Slogger) {
-        let mut rng = StdRng::seed_from_u64(0);
-        let result = match_students(
-            self.members_absent.clone().into_values().collect(),
-            &self
-                .members_present
-                .clone()
-                .into_values()
-                .collect::<Vec<_>>(),
-            &mut rng,
-        );
-
-        for present in self.members_present.values() {
-            for absent in result.placed.get(&present.name).unwrap_or(&Vec::new()) {
-                self.members_unrepresented.remove(&absent.name);
-                self.members_represented
-                    .insert(absent.name.clone(), present.name.clone());
-                debug!(log, "Proxy assigned"; "proxy_for" => absent.name.clone(), "proxied_by" => present.name.clone());
-            }
-        }
-
-        self.members_represented.sort_keys();
-        self.members_unrepresented.sort();
-    }
-}
-
 #[post("/solution", data = "<problem>")]
-fn solution(log: Slogger, problem: Json<ProxyProblem>) -> Value {
+fn post_solution(log: Slogger, problem: Json<ProxyProblem>) -> Value {
     let mut solution = ProxySolution::from_problem(&problem);
 
     info!(log, "Beginning solution"; "metrics" => solution.get_metrics());
@@ -178,7 +48,7 @@ fn rocket() -> _ {
 
     rocket::build()
         .attach(fairing)
-        .mount("/", routes![health_ready, solution])
+        .mount("/", routes![get_health_ready, post_solution])
 }
 
 #[cfg(test)]
@@ -192,7 +62,7 @@ mod test {
     #[test]
     fn test_health_ready() {
         let client = Client::tracked(rocket()).expect("valid rocket instance");
-        let response = client.get(uri!(super::health_ready)).dispatch();
+        let response = client.get(uri!(super::get_health_ready)).dispatch();
         assert_eq!(response.status(), Status::NoContent);
     }
 
@@ -200,7 +70,7 @@ mod test {
     fn test_first_choice_available() {
         let client = Client::tracked(rocket()).expect("valid rocket instance");
         let response = client
-            .post(uri!(super::solution()))
+            .post(uri!(super::post_solution()))
             .header(ContentType::JSON)
             .body(
                 r#"{
@@ -228,7 +98,7 @@ mod test {
     fn test_second_choice_available() {
         let client = Client::tracked(rocket()).expect("valid rocket instance");
         let response = client
-            .post(uri!(super::solution()))
+            .post(uri!(super::post_solution()))
             .header(ContentType::JSON)
             .body(
                 r#"{
@@ -267,7 +137,7 @@ mod test {
 
         let client = Client::tracked(rocket()).expect("valid rocket instance");
         let response = client
-            .post(uri!(super::solution()))
+            .post(uri!(super::post_solution()))
             .header(ContentType::JSON)
             .body(&request)
             .dispatch();
